Add files via upload
Compressed file of the solutions for the tested reactions from the Rhea and MetaNetX databases. This file should be opened as a json dictionary. The first layer is the Rhea or MetaNetX ID, and the second layer contains the inputted reaction SMILES string, along with the outputted MechFind solution. Each solution is a the list of rules found in Unique_Rules.csv